+++ sketch-compose-resources/src/commonMain/kotlin/com/github/panpf/sketch/request/ImageRequest.composeResources.kt
@@ -30,3 +30,15 @@ fun ImageRequest.Builder.error(
     defaultResource: DrawableResource,
     configBlock: @Composable (ErrorStateImage.Builder.() -> Unit)? = null
 ): ImageRequest.Builder = error(ComposableErrorStateImage(defaultResource, configBlock))
+
+
+/**
+ * Set Color image to display when loading fails.
+ *
+ * You can also set image of different error types via the trailing lambda function
+ */
+@Composable
+fun ImageRequest.Builder.composableError(
+    defaultResource: DrawableResource,
+    configBlock: @Composable (ErrorStateImage.Builder.() -> Unit)? = null
+): ImageRequest.Builder = error(ComposableErrorStateImage(defaultResource, configBlock))